#3622 Restrict the v13 conversion to the individual and the population
Formulations, events and compounds are rebuilt from the database templates when a
simulation is created, so their saved structure never reaches the model. Verified
against the provided test project: the saved formulation and compounds carry none of
the new parameters, yet every simulation still runs after converting only the
individual and the population. The event, formulation and compound conversions are
removed, together with the now unused helpers and the inferred requirement to drop the
Meal_StopEvent container, which came from the issue's database-diff table rather than
an acceptance criterion.

Consolidates the v13 converter specs into a single file and moves the v13-only
updaters next to the converter.
