+++ src/PKSim.Infrastructure/ProjectConverter/v13/Converter12To13.cs
@@ -1,59 +1,44 @@
-using System;
-using System.Linq;
 using System.Xml.Linq;
 using OSPSuite.Core.Domain;
-using OSPSuite.Utility.Extensions;
 using OSPSuite.Utility.Visitor;
 using PKSim.Core;
 using PKSim.Core.Model;
-using PKSim.Core.Repositories;
 using PKSim.Core.Services;
 using CoreConverter121To130 = OSPSuite.Core.Converters.v13.Converter121To130;
 
 namespace PKSim.Infrastructure.ProjectConverter.v13
 {
    /// <summary>
-   ///    Brings projects saved before v13 in line with the new oral absorption and particle dissolution model. The model
-   ///    added parameters and containers across most building blocks and changed the definition of the lumen parameters.
-   ///    Everything is cloned from the templates the PK-Sim database provides; nothing is recreated by hand.
+   ///    Brings projects saved before v13 in line with the new oral absorption and particle dissolution model. Only the
+   ///    individual and the population are converted: they are the building blocks whose structure is used directly to
+   ///    build a simulation and whose parameters the user can have edited. Formulations, events and compounds are rebuilt
+   ///    from the database templates when a simulation is created, so their saved structure never reaches the model and
+   ///    does not need to be converted.
    /// </summary>
    public class Converter12To13 : IObjectConverter,
       IVisitor<Individual>,
       IVisitor<Population>,
-      IVisitor<Compound>,
-      IVisitor<Formulation>,
-      IVisitor<PKSimEvent>,
       IVisitor<Simulation>
    {
       private readonly CoreConverter121To130 _coreConverter;
       private readonly IDefaultIndividualRetriever _defaultIndividualRetriever;
-      private readonly IFormulationRepository _formulationRepository;
-      private readonly IEventGroupRepository _eventGroupRepository;
       private readonly ITemplateStructureUpdater _templateStructureUpdater;
       private readonly IIndividualCalculationMethodsUpdater _individualCalculationMethodsUpdater;
       private readonly IPopulationParameterValuesUpdater _populationParameterValuesUpdater;
-      private readonly Lazy<Compound> _templateCompound;
       private bool _converted;
 
       public Converter12To13(
          CoreConverter121To130 coreConverter,
-         ICompoundFactory compoundFactory,
          IDefaultIndividualRetriever defaultIndividualRetriever,
-         IFormulationRepository formulationRepository,
-         IEventGroupRepository eventGroupRepository,
          ITemplateStructureUpdater templateStructureUpdater,
          IIndividualCalculationMethodsUpdater individualCalculationMethodsUpdater,
          IPopulationParameterValuesUpdater populationParameterValuesUpdater)
       {
          _coreConverter = coreConverter;
          _defaultIndividualRetriever = defaultIndividualRetriever;
-         _formulationRepository = formulationRepository;
-         _eventGroupRepository = eventGroupRepository;
          _templateStructureUpdater = templateStructureUpdater;
          _individualCalculationMethodsUpdater = individualCalculationMethodsUpdater;
          _populationParameterValuesUpdater = populationParameterValuesUpdater;
-         //Building a compound from the database is expensive and the template never varies, so it is built at most once
-         _templateCompound = new Lazy<Compound>(compoundFactory.Create);
       }
 
       public bool IsSatisfiedBy(int version) => version == ProjectVersions.V12;
@@ -75,12 +60,6 @@ namespace PKSim.Infrastructure.ProjectConverter.v13
 
       public void Visit(Population population) => convertPopulation(population);
 
-      public void Visit(Compound compound) => convertCompound(compound);
-
-      public void Visit(Formulation formulation) => convertFormulation(formulation);
-
-      public void Visit(PKSimEvent pkSimEvent) => convertEvent(pkSimEvent);
-
       public void Visit(Simulation simulation) => convertSimulation(simulation);
 
       private void convertSimulation(Simulation simulation)
@@ -88,14 +67,11 @@ namespace PKSim.Infrastructure.ProjectConverter.v13
          if (simulation == null)
             return;
 
-         //A simulation keeps its own copy of the building blocks it was created from, so they are converted as well
+         //A simulation keeps its own copy of the individual it was created from, so it is converted as well
          _individualCalculationMethodsUpdater.AddMissingCalculationMethodsTo(simulation);
 
          convertIndividual(simulation.BuildingBlock<Individual>());
          convertPopulation(simulation.BuildingBlock<Population>());
-         simulation.Compounds.Each(convertCompound);
-         simulation.AllBuildingBlocks<Formulation>().Each(convertFormulation);
-         simulation.AllBuildingBlocks<PKSimEvent>().Each(convertEvent);
       }
 
       private void convertPopulation(Population population)
@@ -143,78 +119,6 @@ namespace PKSim.Infrastructure.ProjectConverter.v13
          //The new parameters are species dependent, so the template has to match the population of the individual
          var speciesPopulation = individual.OriginData?.Population;
          return speciesPopulation == null ? null : _defaultIndividualRetriever.DefaultIndividualFor(speciesPopulation);
-      }
-
-      private void convertCompound(Compound compound)
-      {
-         if (compound == null)
-            return;
-
-         var templateCompound = _templateCompound.Value;
-
-         //Only the parameters defined directly in the compound are synchronized here. Its sub containers hold the
-         //parameter alternatives, which are named by the user and must not be aligned with the template by name
-         if (_templateStructureUpdater.AddMissingParametersTo(compound, templateCompound))
-            _converted = true;
-
-         convertCompoundAlternatives(compound, templateCompound);
-      }
-
-      /// <summary>
-      ///    Some of the new compound parameters belong to a group that the user can define alternatives for, solubility
-      ///    among them. Every alternative carries its own copy of the parameters of its group, so each of them has to be
-      ///    brought up to date against the template alternative.
-      /// </summary>
-      private void convertCompoundAlternatives(Compound compound, Compound templateCompound)
-      {
-         foreach (var templateGroup in templateCompound.AllParameterAlternativeGroups())
-         {
-            var templateAlternative = templateGroup.AllAlternatives.FirstOrDefault();
-            var compoundGroup = compound.ParameterAlternativeGroup(templateGroup.Name);
-            if (templateAlternative == null || compoundGroup == null)
-               continue;
-
-            foreach (var alternative in compoundGroup.AllAlternatives)
-            {
-               if (_templateStructureUpdater.AddMissingParametersTo(alternative, templateAlternative))
-                  _converted = true;
-            }
-         }
-      }
-
-      private void convertFormulation(Formulation formulation)
-      {
-         if (formulation == null)
-            return;
-
-         var templateFormulation = _formulationRepository.FormulationBy(formulation.FormulationType);
-         if (_templateStructureUpdater.AddMissingParametersTo(formulation, templateFormulation))
-            _converted = true;
-      }
-
-      private void convertEvent(PKSimEvent pkSimEvent)
-      {
-         if (pkSimEvent == null)
-            return;
-
-         if (removeObsoleteContainersFrom(pkSimEvent))
-            _converted = true;
-
-         var templateEventGroup = _eventGroupRepository.FindByName(pkSimEvent.TemplateName);
-         //The new reset events are whole containers, so the event is synchronized recursively
-         if (_templateStructureUpdater.AddMissingStructureTo(pkSimEvent, templateEventGroup))
-            _converted = true;
-      }
-
-      private static bool removeObsoleteContainersFrom(PKSimEvent pkSimEvent)
-      {
-         var obsoleteContainers = pkSimEvent
-            .GetAllChildren<IContainer>(x => x.IsNamed(ConverterConstants.Containers.MEAL_STOP_EVENT))
-            .ToList();
-
-         obsoleteContainers.Each(x => x.ParentContainer?.RemoveChild(x));
-
-         return obsoleteContainers.Any();
       }
    }
 }
